requirements.txt
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pdfminer.six==20240706
pillow==10.4.0
pycparser==2.22
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
python-dotenv==1.0.1
reportlab==4.2.4
requests==2.32.3
soupsieve==2.6
urllib3==2.2.3
Werkzeug==3.0.4
